chore(lazy-lock.json): update commit hashes for LuaSnip, catppuccin, friendly-snippets, fzf, fzf.vim, gitsigns.nvim, lualine.nvim, mason-lspconfig.nvim, mason.nvim, nerdtree, none-ls.nvim, nui.nvim, nvim-dap, nvim-dap-ui, nvim-lspconfig, nvim-treesitter, nvim-web-devicons, tagbar, vim-commentary, and vim-eunuch
The commit hashes for several dependencies in the lazy-lock.json file have been updated. This ensures that the latest versions of these dependencies are being used in the project. The updated commit hashes provide a reference to the specific versions of the dependencies that were used at the time of the update.